requirements.txt
discord==2.3.2
discord.py==2.4.0
google-ai-generativelanguage==0.6.6